+++ tests/Functional/TestUtil.php
@@ -24,17 +24,10 @@ class TestUtil
     public static function getConnection(): Connection
     {
         static::createDatabase();
-
         $conn = static::getDoctrineConnection();
-
         self::addDbEventSubscribers($conn);
 
         return $conn;
-    }
-
-    public static function getTempConnection(): Connection
-    {
-        return static::getDoctrineConnection('temporary');
     }
 
     public static function getParamsForMainConnection(): array
@@ -59,32 +52,6 @@ class TestUtil
         return $connectionParams;
     }
 
-    public static function getParamsForTemporaryConnection(): array
-    {
-        $connectionParams = [
-            'driver' => $GLOBALS['tmpdb_type'],
-            'user' => $GLOBALS['tmpdb_username'],
-            'password' => $GLOBALS['tmpdb_password'],
-            'host' => $GLOBALS['tmpdb_host'],
-            'dbname' => null,
-            'port' => $GLOBALS['tmpdb_port'],
-        ];
-
-        if (isset($GLOBALS['tmpdb_name'])) {
-            $connectionParams['dbname'] = $GLOBALS['tmpdb_name'];
-        }
-
-        if (isset($GLOBALS['tmpdb_server'])) {
-            $connectionParams['server'] = $GLOBALS['tmpdb_server'];
-        }
-
-        if (isset($GLOBALS['tmpdb_unix_socket'])) {
-            $connectionParams['unix_socket'] = $GLOBALS['tmpdb_unix_socket'];
-        }
-
-        return $connectionParams;
-    }
-
     private static function hasRequiredConnectionParams(): bool
     {
         return isset(
@@ -94,33 +61,20 @@ class TestUtil
             $GLOBALS['db_host'],
             $GLOBALS['db_name'],
             $GLOBALS['db_port']
-        )
-        && isset(
-            $GLOBALS['tmpdb_type'],
-            $GLOBALS['tmpdb_username'],
-            $GLOBALS['tmpdb_password'],
-            $GLOBALS['tmpdb_host'],
-            $GLOBALS['tmpdb_port']
         );
     }
 
     private static function initializeDatabase(): void
     {
         $realConn = static::createDoctrineConnection(self::getParamsForMainConnection());
-        $tmpConn = static::createDoctrineConnection(self::getParamsForTemporaryConnection());
 
-        $platform = $tmpConn->getDatabasePlatform();
+        $platform = $realConn->getDatabasePlatform();
 
         if ($platform->supportsCreateDropDatabase()) {
-            $dbname = $realConn->getDatabase();
+            $realConn->getDatabase();
             $realConn->close();
-
-            $tmpConn->getSchemaManager()->dropAndCreateDatabase($dbname);
-
-            $tmpConn->close();
         } else {
             $sm = $realConn->getSchemaManager();
-
             $schema = $sm->createSchema();
             $stmts = $schema->toDropSql($realConn->getDatabasePlatform());
 
@@ -145,8 +99,7 @@ class TestUtil
 
     private static function getDoctrineConnection(string $name = null): Connection
     {
-        $connection = DB::connection($name);
-        return $connection->getDoctrineConnection();
+        return DB::connection($name)->getDoctrineConnection();
     }
 
     private static function createDoctrineConnection($params): Connection
